Add files via upload
I copied only the code that may have been able to fix the problems I had with CSS, registration and login errors and that worked. 
The problem now is I cant get a registration onto myphpadmin.
I think it might be xampp keeps disabling for MySql. It is on one minute and off the next and not on long enough for me to test and see if there is actually a connection.
What is wrong with my Xampp? I have uninstalled and put different versions in and the same problem.